Fixed shutil error
# Conflicts:
#	dataset_template.py

+++ dataset_template.py
@@ -28,10 +28,11 @@ class Template:
         project_name = self.os.path.splitext(self.os.path.basename(custom_template_file))[0]
         self.work_dir =  self.os.getenv('SCRATCHDIR') + '/' + project_name
         template_file = self.os.path.join(self.work_dir, self.os.path.basename(custom_template_file))
-        
-        #if not self.os.path.isfile(template_file):       # FA 12/18: It should use custom_template_file as this is what will be updated
-        #    self.shutil.copy2(custom_template_file, self.work_dir)
-        self.shutil.copy2(custom_template_file, self.work_dir)
+
+        if not self.os.path.isfile(template_file):       # FA 12/18: It should use custom_template_file as this is what will be updated
+           self.shutil.copy2(custom_template_file, self.work_dir)
+        #shutil.copy2(custom_template_file, self.work_dir)
+
         self.options = self.read_options(template_file)
